requirements.txt
<<<<<<< HEAD
Flask==3.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
=======
Flask==3.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
>>>>>>> 0647f0198b46afd5b1f82659651058e63626919e
gunicorn==21.2.0